requirements.txt
# Azure
azure-functions==1.21.0
azure-identity==1.17.1
azure-cosmos==4.7.0
azure-servicebus==7.12.2
azure-keyvault-secrets==4.8.0

# Microsoft Graph auth
msal==1.29.0

# HTTP client
httpx==0.27.0

# Retry logic
tenacity==8.3.0

# AI / Agent framework
openai==1.40.0
langchain==0.2.12
langchain-openai==0.1.20
langgraph==0.2.4

# Data validation
pydantic==2.8.2

# Local dev
python-dotenv==1.0.1

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0

# For prompts in ai_engine folder
jinja2==3.1.4

# Frontend
streamlit==1.35.0